python-llm/requirements.txt
# Core dependencies
python-dotenv
composio
composio-gemini
google-genai